+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - error-headers option
 - ignore-missing-headers option
+- ignore-ssl-errors option
 
 ## 0.5.0 - 2020-08-11
 ### Added